Use openmpi-4.0.1 built with the default gcc compiler
This build of OpenMPI is built with gcc-7.4.0, which is the default
system compiler. This avoids issues with some packages, such as
ScaLAPACK, that rely on cmake and fail to build due to using an
incompatible version of gfortran. The previous OpenMPI module was
built after loading a module for gcc-8.2.0.