Remove 'gcc' from modules paths and names
This was intended to distinguish between modules built with different
toolchains. However, in practice, the current modules are only built
with the same toolchain.

A better scheme for distinguishing between modules built with
different toolchains could be to instead add the name of the toolchain
as a suffix. For example, 'openmpi/4.0.1' would be the default version
of OpenMPI, usually built with the system compiler, whereas
'openmpi/4.0.1-intel' would be OpenMPI built with the default Intel
toolchain, and 'openmpi-intel-2019-19.04' would be OpenMPI built
specifically with the Intel Compiler version 2019-19.04.