several fixes related to python packages, e.g., matplotlib and scipy
Update scipy to 1.8.1 (since 1.7.3 did not build on arm).

Add python-beniget-0.4.1, which is needed as a dependency.

Fix PYTHONPATH for several packages that were installed with `pip
install --target=DIR' instead of `pip install --prefix=DIR'. The
former installs directly to the specified directory, whereas the
latter installs the packages into a site-packages directory.